+++ R/create_neighbourhood.R
@@ -16,16 +16,16 @@
 #' @return matrix
 #'
 #' @examples
+#' \dontrun{
 #' mat <- matrix(1, nrow= 10, ncol = 10)
 #' cell_id <- rbind(cbind(3,5), cbind(7,1))
 #' create_neighbourhood(cell_id, mat)
+#' }
 #'
 #' @aliases create_neighbourhood
 #' @rdname create_neighbourhood
 #'
 #' @keywords internal
-
-#' @export
 create_neighbourhood <- function(cells, matrix, directions = 4) {
 
   if (directions == 4) {